chore(*): Refactor to re-use common code for header sizing from grid base.
